backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.0.2
tweepy==4.10.0
pillow==8.3.1
transformers==4.11.3